linux-odroid-600: Copy 5.18 config and run make oldconfig
For prompts I referred to the current (6.4) kernel config in Arch Linux
or picked the default, except for LSM.

SELinux and AppArmor are useful for a lot of modern applications (like
Docker or Kubernetes). Smack and TOMOYO probably less so, but they were
enable before, so we should keep them for compatibility.

https://www.kernel.org/doc/html/latest/admin-guide/LSM/tomoyo.html
https://www.kernel.org/doc/html/latest/admin-guide/LSM/Smack.html